% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivr.R
\name{archiv.fromText}
\alias{archiv.fromText}
\title{Save the links in a text file (docx, pdf, markdown) in perma.cc or Wayback.}
\usage{
archiv.fromText(fp, method = "wayback")
}
\arguments{
\item{fp}{The filepath to extract links from.}

\item{method}{Either "wayback," "perma_cc" or "both".}
}
\value{
a dataframe containing the url, status, availability,
  archived url(s) and timestamp(s)
}
\description{
Save the links in a text file (docx, pdf, markdown) in perma.cc or Wayback.
}
