% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivr.R
\name{archiv_batch}
\alias{archiv_batch}
\title{Save a batch of urls to a folder - THIS CURRENTLY DOES NOT WORK.}
\usage{
archiv_batch(url_list, api = .perma_cc_key, folder = .folder_id)
}
\arguments{
\item{url_list}{A vector of urls to archive.}

\item{api}{(Optional api key)}

\item{folder}{(Mandatory, but defaults to .folder_id)}
}
\description{
Save a batch of urls to a folder - THIS CURRENTLY DOES NOT WORK.
}
