% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivr.R
\name{archiv_perma}
\alias{archiv_perma}
\title{Saves a single url in perma.cc.}
\usage{
archiv_perma(arc_url, method = "perma_cc")
}
\arguments{
\item{arc_url}{The url to archive.}
}
\value{
A list or object representing the result.
}
\description{
Saves a single url in perma.cc.
}
