% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivr.R
\name{archiv_url}
\alias{archiv_url}
\title{Saves a single url in either perma.cc or the wayback machine.}
\usage{
archiv_url(arc_url, method = "perma_cc")
}
\arguments{
\item{arc_url}{The url to archive.}

\item{method}{Either "perma_cc" or the default, "wayback."}
}
\value{
A list or object representing the result.
}
\description{
Saves a single url in either perma.cc or the wayback machine.
}
