% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivr.R
\docType{package}
\name{archivr}
\alias{archivr}
\alias{archivr-package}
\title{Copyright <2019> <Qualitative Data Repository, Syracuse University>
Permission is hereby granted, free of charge, to any person obtaining a copy
of this software and associated documentation files (the "Software"), to deal
in the Software without restriction, including without limitation the rights
to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
copies of the Software, and to permit persons to whom the Software is
furnished to do so, subject to the following conditions:
The above copyright notice and this permission notice shall be included in
all copies or substantial portions of the Software.
THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
THE SOFTWARE.
Archivr: Save Your Websites in Perma.cc or the Wayback Machine}
\description{
Archivr is a toolkit for the long-run archiving of Qualitative data.
It takes a list of urls and uses either the perma.cc or Wayback Machine
archives to store the webpages for future reference. It will also parse
word or html documents for urls to be archived.
}
