% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_folder}
\alias{.check_folder}
\title{Works with get_subfolders to flatten the folder ids tree}
\usage{
.check_folder(folder_list)
}
\arguments{
\item{folder_list}{a list of perma.cc folder objects}
}
\value{
A list of vectors with the id and name.
}
\description{
Works with get_subfolders to flatten the folder ids tree
}
