% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_subfolders}
\alias{.get_subfolders}
\title{Works with check_folder to flatten folder ids tree}
\usage{
.get_subfolders(id)
}
\arguments{
\item{id}{A folder id}
}
\value{
A list of vectors with the id and name.
}
\description{
Works with check_folder to flatten folder ids tree
}
