% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivr.R
\docType{data}
\name{.perma_cc_folder_pref}
\alias{.perma_cc_folder_pref}
\title{Global var for the API key for perma.cc}
\format{An object of class \code{character} of length 1.}
\usage{
.perma_cc_folder_pref
}
\description{
Global var for the API key for perma.cc
}
\keyword{datasets}
