% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apitools.R
\docType{data}
\name{.wb_available_url}
\alias{.wb_available_url}
\title{Default url for the Wayback Machine}
\format{An object of class \code{character} of length 1.}
\usage{
.wb_available_url
}
\description{
Default url for the Wayback Machine
}
\keyword{datasets}
