% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivr.R
\name{extract_urls_from_text}
\alias{extract_urls_from_text}
\title{Get the urls from a text file or string}
\usage{
extract_urls_from_text(fp)
}
\arguments{
\item{fp}{A filepath or string.}
}
\value{
a List of Urls.
}
\description{
Get the urls from a text file or string
}
