% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivr.R
\name{extract_urls_from_webpage}
\alias{extract_urls_from_webpage}
\title{Extracts the urls from a webpage.}
\usage{
extract_urls_from_webpage(url)
}
\arguments{
\item{url}{The url to extract urls.}
}
\value{
a vector of urls.
}
\description{
Extracts the urls from a webpage.
}
