% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivr.R
\name{from_wayback}
\alias{from_wayback}
\title{Check whether a url is available in the Wayback Machine}
\usage{
from_wayback(url)
}
\arguments{
\item{url}{The url to check.}
}
\value{
a list containing
  the original url.
  TRUE if successful or FALSE
  the archived url.
  the last time the url was crawled.
}
\description{
Check whether a url is available in the Wayback Machine
}
\examples{
checkArchiveStatus <- from_wayback(
    "https://www-cs-faculty.stanford.edu/~knuth/retd.html"
    )
}
