% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivr.R
\name{get_api_key}
\alias{get_api_key}
\title{Get the perma.cc api key if set.}
\usage{
get_api_key()
}
\value{
The current api key state.
}
\description{
Get the perma.cc api key if set.
}
\examples{
set_api_key("API KEY")
get_api_key()
}
