% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivr.R
\name{get_default_folder}
\alias{get_default_folder}
\title{Get the folder id and name from all text files in a perma.cc folder}
\usage{
get_default_folder(default = 1)
}
\arguments{
\item{default}{The value of the folder to collect by default.}
}
\value{
The id and name of the first top folder (usually "Personal Links")
   in perma.cc
}
\description{
Get the folder id and name from all text files in a perma.cc folder
}
