% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivr.R
\name{get_folder_ids}
\alias{get_folder_ids}
\title{Get the folder ids starting from the default folder.}
\usage{
get_folder_ids()
}
\value{
A list of vectors with the top folder and all its children.
}
\description{
Get the folder ids starting from the default folder.
}
