% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivr.R
\name{list_string}
\alias{list_string}
\title{Creates a json string from a list of urls.}
\usage{
list_string(url_list)
}
\arguments{
\item{url_list}{A list of urls.}
}
\value{
A json string representing the list.
}
\description{
Creates a json string from a list of urls.
}
\examples{
urls <- c("https://qdr.syr.edu", "https://cran.r-project.org/", "https://apsa.net")
list_string(urls)
}
