% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_archiv.R
\name{memento_check}
\alias{memento_check}
\title{Check whether a url is available from archive.md via memento}
\usage{
memento_check(url)
}
\arguments{
\item{url}{The url to check.}
}
\value{
a list containing
  the original url.
  Whether it's available
  the archived url, if available.
}
\description{
Check whether a url is available from archive.md via memento
}
\examples{
\dontrun{memento_check("https://google.com")}
}
