% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivr.R
\name{set_api_key}
\alias{set_api_key}
\title{Set the api key(s) for Perma.cc apis, if required.}
\usage{
set_api_key(key)
}
\arguments{
\item{key}{The Api Key.}
}
\description{
Set the api key(s) for Perma.cc apis, if required.
}
\examples{
set_api_key("API_KEY")
}
