% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivr.R
\name{set_folder_id}
\alias{set_folder_id}
\title{Set the folder to save items in Perma.cc.}
\usage{
set_folder_id(id)
}
\arguments{
\item{id}{The folder id. This will be a string of numbers. If you do not
know your folder id, get_folder_ids() will output a complete list of
folders}
}
\value{
TRUE
}
\description{
Set the folder to save items in Perma.cc.
}
\examples{
set_folder_id("42")
}
\seealso{
[get_folder_ids()]
}
