% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivr.R
\name{view_archiv.fromUrl}
\alias{view_archiv.fromUrl}
\title{Collect information on whether links in a url are archived.}
\usage{
view_archiv.fromUrl(url, method = "wayback")
}
\arguments{
\item{url}{The url to extract links from.}

\item{method}{Either "wayback," "perma_cc" or "both".}
}
\value{
a dataframe containing the url, status, availability,
  archived url(s) and timestamp(s)
}
\description{
Collect information on whether links in a url are archived.
}
