% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivr.R
\name{view_archiv}
\alias{view_archiv}
\title{Get archiving data from a list of Urls}
\usage{
view_archiv(lst, method = "wayback")
}
\arguments{
\item{lst}{A list of urls to check.}

\item{method}{"wayback", "perma_cc" or "both".}
}
\value{
A dataframe containing the original urls, their http status,
 availability, the archive url if it exists and a timestamp for the last
 web crawl.
}
\description{
Get archiving data from a list of Urls
}
