% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_archiv.R
\name{view_perma_cc}
\alias{view_perma_cc}
\title{Check whether a url is available in Perma.cc}
\usage{
view_perma_cc(url)
}
\arguments{
\item{url}{The url to check.}
}
\value{
a list containing
  the original url.
  TRUE if successful or FALSE
  the archived url.
  the last time the url was crawled.
}
\description{
Check whether a url is available in Perma.cc
}
\examples{
checkArchiveStatus <- view_perma_cc(
    "https://www-cs-faculty.stanford.edu/~knuth/retd.html"
    )
}
