% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivr.R
\name{archiv.fromText}
\alias{archiv.fromText}
\title{Collect information on whether links in a file are archived.}
\usage{
archiv.fromText(fp, method = "wayback")
}
\arguments{
\item{fp}{The filepath to extract links from.}

\item{method}{Either "wayback," "perma_cc" or "both".}
}
\value{
a dataframe containing the url, status, availability,
  archived url(s) and timestamp(s)
}
\description{
Collect information on whether links in a file are archived.
}
