% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivr.R
\name{archiv.fromUrl}
\alias{archiv.fromUrl}
\title{Save the links in a url in perma.cc or Wayback.}
\usage{
archiv.fromUrl(url, method = "wayback", except = NULL)
}
\arguments{
\item{url}{The url to extract links from.}

\item{method}{Either "wayback," "perma_cc" or "both".}

\item{except}{A regular expression for URLs to exclude from extraction}
}
\value{
a dataframe containing the url, availability,
  archived url(s) and timestamp(s)
}
\description{
Save the links in a url in perma.cc or Wayback.
}
\examples{
# Wayback
archivedURLs <- archiv.fromUrl(
     "https://www-cs-faculty.stanford.edu/~knuth/retd.html",
     except="validator\\\\.w3\\\\.org"
     )

#perma.cc
\dontrun{
set_api_key("API KEY")
set_folder_id("42")
archivedURLs <- archiv.fromUrl(
     "https://www-cs-faculty.stanford.edu/~knuth/retd.html",
     method="perma_cc"
     )
}
}
