% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivr.R
\name{archiv.fromUrl}
\alias{archiv.fromUrl}
\title{Save the links in a url in perma.cc or Wayback.}
\usage{
archiv.fromUrl(url, method = "wayback")
}
\arguments{
\item{url}{The url to extract links from.}

\item{method}{Either "wayback," "perma_cc" or "both".}
}
\value{
a dataframe containing the url, status, availability,
  archived url(s) and timestamp(s)
}
\description{
Save the links in a url in perma.cc or Wayback.
}
\examples{
# Wayback
archiv.fromUrl("https://www-cs-faculty.stanford.edu/~knuth/retd.html")

#perma.cc
\dontrun{
set_api_key("API KEY")
set_folder_id("42")
archiv.fromUrl("https://www-cs-faculty.stanford.edu/~knuth/retd.html", method="perma_cc")
}
}
