% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivr.R
\name{archiv}
\alias{archiv}
\title{Archive a list of urls in Wayback or perma_cc.}
\usage{
archiv(url_list, method = "wayback")
}
\arguments{
\item{url_list}{A list of urls to archive.}

\item{method}{Either "wayback" or "perma_cc." Defaults to "wayback."}
}
\value{
A dataframe containing the original urls, the urls to the
  archived website. For Perma.cc also the URL to the screenshot, the short URL and a timestamp.
}
\description{
Archive a list of urls in Wayback or perma_cc.
}
