% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivr.R
\name{archiv}
\alias{archiv}
\title{Archive a list of urls in Wayback or perma_cc.}
\usage{
archiv(url_list, method = "wayback")
}
\arguments{
\item{url_list}{A list of urls to archive.}

\item{method}{Either "wayback" or "perma_cc." Defaults to "wayback."}
}
\value{
A dataframe containing the original urls, the urls to the
  archived website and a timestamp. For Perma.cc also the URL to the
  screenshot, and the short URL.
}
\description{
Archive a list of urls in Wayback or perma_cc.
}
\examples{
urls <- c("https://qdr.syr.edu", "https://cran.r-project.org/", "https://apsa.net")

# archive in Wayback machine
archivedUrls <- archiv(urls)

# archive in perma.cc
\dontrun{
set_api_key("API KEY")
set_folder_id("FOLDER ID")
archivedUrls <- archiv(urls, method="perma_cc")
}

}
