% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivr.R
\name{archiv_wayback}
\alias{archiv_wayback}
\title{Save a url on the wayback machine.}
\usage{
archiv_wayback(arc_url)
}
\arguments{
\item{arc_url}{- the url to archive.}
}
\value{
A list or object representing the result.
}
\description{
Save a url on the wayback machine.
}
\examples{
archiv_wayback("https://qdr.syr.edu")
}
