% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archiv.R
\name{archiv_wayback}
\alias{archiv_wayback}
\title{Save a url on the wayback machine.}
\usage{
archiv_wayback(arc_url)
}
\value{
A list representing the result.
}
\description{
Save a url on the wayback machine.
}
\examples{
\dontrun{
archivedUrl <- archiv_wayback("https://qdr.syr.edu")
}
}
