% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivr.R
\docType{package}
\name{archivr}
\alias{archivr}
\alias{archivr-package}
\title{Archivr: Save Your Websites in Perma.cc or the Wayback Machine}
\description{
Archivr is a toolkit for the long-run archiving of qualitative data.
It takes a list of urls and uses either the perma.cc or Wayback Machine
archives to store the webpages for future reference. It will also parse
documents or webpages for urls to be archived.
}
