% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivr.R
\name{extract_urls_from_folder}
\alias{extract_urls_from_folder}
\title{Get the urls from all text, pdf, or docx files in a folder}
\usage{
extract_urls_from_folder(fp)
}
\arguments{
\item{fp}{A filepath or string.}
}
\value{
A list of urls.
}
\description{
Get the urls from all text, pdf, or docx files in a folder
}
