% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_urls.R
\name{extract_urls_from_webpage}
\alias{extract_urls_from_webpage}
\title{Extracts the urls from a webpage.}
\usage{
extract_urls_from_webpage(url, except = NULL)
}
\arguments{
\item{url}{The url to extract urls.}

\item{except}{A regular expression for URLs to exclude from extraction}
}
\value{
a vector of urls.
}
\description{
The function works simply by extracting the `href`` attribute from all `a` nodes.
It is called internally from `archiv.fromUrl` but can be useful as a separate function
if you want to filter which links you archive.
}
\examples{
urlList <- extract_urls_from_webpage(
     "https://www-cs-faculty.stanford.edu/~knuth/retd.html",
     except="validator\\\\.w3\\\\.org"
     )
}
