% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivr.R
\name{from_perma_cc}
\alias{from_perma_cc}
\title{Check whether a url is available in Perma.cc}
\usage{
from_perma_cc(url)
}
\arguments{
\item{url}{The url to check.}
}
\value{
a vector containing
  the original url.
  the http status
  TRUE if successful or FALSE
  the archived url.
  the last time the url was crawled.
}
\description{
Check whether a url is available in Perma.cc
}
\examples{
from_perma("https://www-cs-faculty.stanford.edu/~knuth/retd.html")
}
