% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apitools.R
\name{get_default_folder}
\alias{get_default_folder}
\title{Get the folder id of the perma.cc default folder (usually "Personal Links")}
\usage{
get_default_folder(default = 1)
}
\arguments{
\item{default}{The number of the folder to accept by default.}
}
\value{
The id and name of the first top folder (usually "Personal Links")
   in perma.cc
}
\description{
Get the folder id of the perma.cc default folder (usually "Personal Links")
}
\examples{
\dontrun{
set_api_key("API_KEY")
permaFolder <- get_default_folder()
}
}
