% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivr.R
\name{get_folder_id}
\alias{get_folder_id}
\title{Get the root folder id for the current api key.}
\usage{
get_folder_id()
}
\value{
The current folder id state.
}
\description{
Get the root folder id for the current api key.
}
