% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivr.R
\name{view_archiv.fromUrl}
\alias{view_archiv.fromUrl}
\title{Collect information on whether links contained in a webpage are archived.}
\usage{
view_archiv.fromUrl(url, method = "wayback")
}
\arguments{
\item{url}{The url of the webpage to extract links from.}

\item{method}{Either "wayback," "perma_cc" or "both".}
}
\value{
a dataframe containing the url, status, availability,
  archived url(s) and timestamp(s)
}
\description{
Collect information on whether links contained in a webpage are archived.
}
\examples{
checkArchiveStatus <- view_archiv.fromUrl("https://www-cs-faculty.stanford.edu/~knuth/retd.html", method="both")
}
