% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivr.R
\name{view_archiv}
\alias{view_archiv}
\title{Get archiving data from a list of Urls}
\usage{
view_archiv(lst, method = "wayback")
}
\arguments{
\item{lst}{A list of urls to check.}

\item{method}{"wayback", "perma_cc" or "both".}
}
\value{
A dataframe containing the original urls,
 availability, the archive url if it exists and a timestamp for the last
 web crawl. 
 
 Where method is "both", "availability" is TRUE if the URL is archived by either
 service
}
\description{
Get archiving data from a list of Urls
}
\examples{
urls <- c("https://qdr.syr.edu", "https://cran.r-project.org/", "https://apsa.net")
checkArchiveStatus <- view_archiv(urls, method="both")
}
